+++ src/components/Dropdown/index.js
@@ -4,55 +4,12 @@ import ArrowDownIcon from './ArrowDownIcon'
 import config from 'vtex-tachyons/config.json'
 
 class Dropdown extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      open: false,
-      optionsWidth: 0,
-    }
+  handleChange = e => {
+    const { disabled, onChange } = this.props
+    const { target: { value } } = e
+
+    !disabled && onChange && onChange(e, value)
   }
-
-  componentDidMount() {
-    document.addEventListener('mousedown', this.handleClickOutside)
-    this.setState({ optionsWidth: this.wrapperRef.clientWidth })
-  }
-
-  componentWillUnmount() {
-    document.removeEventListener('mousedown', this.handleClickOutside)
-  }
-
-  setWrapperRef = node => {
-    this.wrapperRef = node
-  };
-
-  setSelectRef = el => {
-    this.select = el
-  };
-
-  handleClickOutside = e => {
-    if (this.wrapperRef && !this.wrapperRef.contains(e.target)) {
-      this.setState({ open: false }, () => {
-        this.props.onClose && this.props.onClose(e)
-      })
-    }
-  };
-
-  handleClick = e => {
-    const { onOpen, onClose } = this.props
-    this.select && this.select.blur()
-    this.setState({ open: !this.state.open }, () => {
-      this.state.open ? onOpen && onOpen(e) : onClose && onClose(e)
-    })
-  };
-
-  handleOptionClick = (e, option) => {
-    const { disabled, onChange, onClose } = this.props
-
-    !disabled && onChange && onChange(e, option)
-    this.setState({ open: false }, () => {
-      onClose && onClose(e)
-    })
-  };
 
   getValueLabel() {
     const option = this.props.options.find(
@@ -66,56 +23,54 @@ class Dropdown extends Component {
     const {
       label,
       id,
+      value,
       size,
       disabled,
       options,
       error,
       errorMessage,
       helpText,
+      placeholder,
+      preventTruncate,
+      autoFocus,
+      form,
+      name,
+      required,
     } = this.props
-    const { open } = this.state
 
     let width
-    let maxHeight
     let iconSize
 
     let classes = 'bg-transparent bn w-100 '
-    let containerClasses = 'br2 bw1 '
-    let optionsClasses = 'absolute bl br bb bw1 br2 br--bottom bg-white flex-column z-max overflow-y-auto '
-    let optionClasses = 'w-100 pointer flex bg-white hover-bg-near-white near-black tl bb-0 bl-0 br-0 bt b--near-white '
+    let containerClasses = 'br2 bw1 relative '
+    let selectClasses = 'o-0 absolute top-0 left-0 w-100 bottom-0 '
 
     const valueLabel = this.getValueLabel()
     const showCaption = !valueLabel
 
     classes += disabled ? 'bg-light-gray ' : 'pointer '
+    selectClasses += disabled ? '' : 'pointer '
     classes += !disabled && valueLabel ? 'near-black ' : 'gray '
 
     switch (size) {
       case 'large':
         classes += 'f5 pv4 pl6 pr5 '
-        optionClasses += 'f5 pv4 ph6 '
-        maxHeight = '200px'
+        selectClasses += 'f5 '
         iconSize = 18
         break
       case 'x-large':
         classes += 'f4 pv5 pl7 pr6 '
-        optionClasses += 'f4 pv5 ph7 '
-        maxHeight = '260px'
+        selectClasses += 'f4 '
         iconSize = 22
         break
       default:
         classes += 'f6 pv3 pl5 pr4 '
-        optionClasses += 'f6 pv3 ph5 '
-        maxHeight = '150px'
+        selectClasses += 'f6 '
         iconSize = 16
         break
     }
 
     const containerStyle = { width }
-    const optionsStyle = {
-      maxHeight: maxHeight,
-      width: this.state.optionsWidth,
-    }
 
     if (disabled) {
       containerClasses += 'bg-light-gray '
@@ -123,41 +78,34 @@ class Dropdown extends Component {
       containerClasses += 'bg-white '
     }
 
-    if (open) {
-      containerClasses += 'bl br bt pb1 b--silver br--top '
-      optionsClasses += 'b--silver '
+    if (error || errorMessage) {
+      containerClasses += 'ba b--red hover-b--red '
     } else {
-      if (error || errorMessage) {
-        containerClasses += 'ba b--red hover-b--red '
-      } else {
-        containerClasses += 'ba b--light-gray '
-      }
-      optionsClasses += 'pointer b--light-gray'
-      if (!disabled) {
-        containerClasses += 'hover-b--silver '
-      }
+      containerClasses += 'ba b--light-gray '
+    }
+
+    if (!disabled) {
+      containerClasses += 'hover-b--silver '
     }
 
     return (
-      <div className="vtex-dropdown" ref={this.setWrapperRef}>
+      <div className="vtex-dropdown">
         <label>
-          {label &&
+          {label && (
             <span className="vtex-dropdown__label dib mb3 w-100">
               {label}
-            </span>}
+            </span>
+          )}
           <div className={containerClasses} style={containerStyle}>
-            <button
+            <div
               id={id}
-              disabled={disabled}
-              ref={this.setSelectRef}
-              onClick={this.handleClick}
               className={`vtex-dropdown__button ${classes}`}
             >
               <div className="flex">
-                <div className="vtex-dropdown__caption flex-auto tl">
-                  {showCaption ? this.props.optionsCaption : valueLabel}
+                <div className="vtex-dropdown__caption flex-auto tl truncate">
+                  {showCaption ? placeholder : valueLabel}
                 </div>
-                <div className="vtex-dropdown__arrow flex-none flex items-center pl6">
+                <div className="vtex-dropdown__arrow flex-none flex items-center pl2">
                   <ArrowDownIcon
                     size={iconSize}
                     color={
@@ -166,24 +114,40 @@ class Dropdown extends Component {
                   />
                 </div>
               </div>
-            </button>
+            </div>
+
+            <select
+              disabled={disabled}
+              className={selectClasses}
+              onChange={this.handleChange}
+              value={value}
+              autoFocus={autoFocus}
+              form={form}
+              name={name}
+              required={required}
+              style={{
+                // safari select height fix
+                webkitAppearance: 'menulist-button',
+              }}
+            >
+              {preventTruncate && (
+                <optgroup label={label || helpText || placeholder || ''}></optgroup>
+              )}
+              {options.map(option => (
+                <option key={option.value} value={option.value}>
+                  {option.label}
+                </option>
+              ))}
+            </select>
+
           </div>
         </label>
-        {open &&
-          <div className={`vtex-dropdown__options ${optionsClasses}`} style={optionsStyle}>
-            {options.map(option => (
-              <button
-                key={option.value}
-                className={optionClasses}
-                onClick={e => this.handleOptionClick(e, option)}
-              >
-                {option.label}
-              </button>
-            ))}
-          </div>}
-        {errorMessage &&
-          <div className="red f6 mt3 lh-title">{errorMessage}</div>}
-        {helpText && <div className="mid-gray f6 mt3 lh-title">{helpText}</div>}
+        {errorMessage && (
+          <div className="red f6 mt3 lh-title">{errorMessage}</div>
+        )}
+        {helpText && (
+          <div className="mid-gray f6 mt3 lh-title">{helpText}</div>
+        )}
       </div>
     )
   }
@@ -191,17 +155,20 @@ class Dropdown extends Component {
 
 Dropdown.defaultProps = {
   size: 'regular',
+  options: [],
 }
 
 Dropdown.propTypes = {
   /** Error highlight */
   error: PropTypes.bool,
   /** Error message */
-  errorMessage: PropTypes.string,
+  errorMessage: PropTypes.oneOfType([PropTypes.node, PropTypes.string]),
   /** Help text */
-  helpText: PropTypes.node,
+  helpText: PropTypes.oneOfType([PropTypes.node, PropTypes.string]),
   /** Dropdown label */
   label: PropTypes.string,
+  /** Dropdown placeholder value */
+  placeholder: PropTypes.string,
   /** Dropdown size */
   size: PropTypes.oneOf(['regular', 'large', 'x-large']),
   /** Dropdown options list */
@@ -211,14 +178,22 @@ Dropdown.propTypes = {
       label: PropTypes.string.isRequired,
     })
   ),
-  /** Dropdown placeholder value */
-  optionsCaption: PropTypes.string,
+  /** Prevent truncating large options texts on some devices/browsers, such as iOS */
+  preventTruncate: PropTypes.bool,
   /** Spec attribute */
   id: PropTypes.string,
+  /** Spec attribute */
+  autoFocus: PropTypes.bool,
   /** Spec attribute */
   value: PropTypes.string,
   /** Spec attribute */
   disabled: PropTypes.bool,
+  /** Spec attribute */
+  form: PropTypes.string,
+  /** Spec attribute */
+  name: PropTypes.string,
+  /** Spec attribute */
+  required: PropTypes.bool,
   /** onChange event */
   onChange: PropTypes.func,
   /** onClose event */